moved src files to src folder
